Fixed Null Pointer exception
There was an unsafe assignment @ ex line 1935 where the result of the call to visit method was directly passed to the string buildere
without checking for it's possibility to be null. The fix is as easy as it can be, it reports the null value and set it as an empty string, 
the reson behind it's null value should be looked into it.